requirements.txt
fastapi
uvicorn
python-dotenv
openai
sse-starlette
httpx
